Исправление недоработок. Закрыта дыра с приватностью в Items

+++ src/main/java/org/example/todolistv2/services/ItemService.java
@@ -2,32 +2,46 @@ package org.example.todolistv2.services;
 
 import org.example.todolistv2.entity.Item;
 import org.example.todolistv2.exceptions.BadRequestException;
+import org.example.todolistv2.exceptions.NoAccessException;
 import org.example.todolistv2.exceptions.NotFoundObjectException;
 import org.example.todolistv2.exceptions.NotFoundOwnerException;
 import org.example.todolistv2.mongotemplates.ItemRepository;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
 
 import java.util.List;
 
+@Service
 public class ItemService {
-    ItemRepository itemRepository;
-    GroupService groupServices;
+    @Autowired
+    private ItemRepository itemRepository;
+    @Autowired
+    private GroupService groupServices;
 
-    public Item create(Item newItem, String groupId) {
+    public Item create(String userId, String groupId ,Item newItem) {
         if (groupServices.exist(groupId)) {
             throw new NotFoundOwnerException();
         }
-        if (newItem != null) {
+        if (newItem.getGroupId() == null || newItem.getName() == null) {
             throw new BadRequestException();
+        }
+        if (newItem.getActivity() == null) {
+            newItem.setActivity(true);
+        }
+        if(groupServices.hasAccess(userId, groupId) || !newItem.getGroupId().equals(groupId)){
+            throw  new NoAccessException();
         }
         return newItem;
     }
 
-    public Item remove(String itemId) {
+    public Item remove(String  userId, String groupId, String itemId) {
         Item removingItem = itemRepository.findItemById(itemId);
         if (removingItem == null) {
             throw new NotFoundObjectException();
+        }
+        if(groupServices.hasAccess(userId, groupId) || !removingItem.getGroupId().equals(groupId)){
+            throw  new NoAccessException();
         }
         itemRepository.delete(removingItem);
         return removingItem;
@@ -43,26 +57,32 @@ public class ItemService {
     }
 
 
-    public List<Item> found(String group_id) {
-        List<Item> itemList = itemRepository.findItemsByGroupId(group_id);
+    public List<Item> found(String userId, String groupId) {
+        List<Item> itemList = itemRepository.findItemsByGroupId(groupId);
         if (itemList == null) {
             throw new NotFoundObjectException();
+        }
+        if(groupServices.hasAccess(userId, groupId)){
+            throw  new NoAccessException();
         }
         return itemList;
     }
 
-    public Item getInfo(String group_id, String item_id) {
-        Item itemInfo = itemRepository.findItemById(item_id);
-        if (groupServices.exist(group_id)) {
+    public Item getInfo(String userId, String groupId, String item_id) {
+        if (groupServices.exist(groupId)) {
             throw new NotFoundOwnerException();
         }
+        Item itemInfo = itemRepository.findItemById(item_id);
         if (itemInfo == null) {
             throw new NotFoundObjectException();
+        }
+        if(groupServices.hasAccess(userId, groupId) || !itemInfo.getGroupId().equals(groupId)){
+            throw  new NoAccessException();
         }
         return itemInfo;
     }
 
-    public Item update(String itemId, String groupId, Item updItem) {
+    public Item update(String userId, String groupId, String itemId, Item updItem) {
         if (updItem == null) {
             throw new BadRequestException();
         }
@@ -70,7 +90,10 @@ public class ItemService {
         if (oldItem == null) {
             throw new NotFoundObjectException();
         }
-        if (updItem.getGroupId() != null && groupId != updItem.getGroupId()) {
+        if(groupServices.hasAccess(userId, groupId)){
+            throw  new NoAccessException();
+        }
+        if (updItem.getGroupId() != null && !groupId.equals(updItem.getGroupId())) {
             if (groupServices.exist(updItem.getGroupId())) {
                 oldItem.setGroupId(updItem.getGroupId());
             } else {
@@ -85,7 +108,6 @@ public class ItemService {
         }
         return oldItem;
     }
-
 
     public void removeByGroupId(String groupId) {
         List<Item> removeItem = itemRepository.findItemsByGroupId(groupId);
